First release
This project was a for of the PortForward, however both projects
did diverge on aim, so I decided to create a new repository.

Message from the previous commit:

Revamp of the code to use YAML instead of JSON for easy maintenance

Remove some 'features' like Quota and Expiration as they seem
outside of the scode of the project.
The saving of file every X seconds is also removed, this should
reduce significantly the overhead of the application running.